requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
websockets>=10.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.20.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Core Data Science
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
joblib>=1.1.0
lightgbm>=3.3.0
xgboost>=1.5.0
catboost>=1.0.0
optuna>=2.10.0
shap>=0.40.0

# GPU Acceleration
cudf-cu11>=22.12.0; platform_system != "Windows"
cuml-cu11>=22.12.0; platform_system != "Windows"
tensorflow>=2.8.0
torch>=1.10.0
scikit-learn-intelex>=2021.5.0

# Deep Learning
keras>=2.8.0

# API and Web
requests>=2.26.0
flask>=2.0.0

# Utilities
tqdm>=4.62.0
ray>=1.13.0
hyperopt>=0.2.7

# Additional dependencies
psutil>=5.8.0
schedule>=1.2.0

# Event-driven trading system dependencies

# Core Python packages
asyncio>=3.4.3
typing-extensions>=4.1.1

# Data handling
pandas>=1.3.5
numpy>=1.21.5

# WebSocket (for real implementations; not used in examples)
websockets>=10.3
aiohttp>=3.8.1

# Utilities
python-dotenv>=0.20.0
pydantic>=1.9.1